Expression visitor moved to separate file
