+++ Driver/Driver.csproj
@@ -74,6 +74,7 @@
       <Link>Properties\GlobalAssemblyInfo.cs</Link>
     </Compile>
     <Compile Include="Builders\BuilderBase.cs" />
+    <Compile Include="Builders\TypedBuilder\ExpressionVisitor.cs" />
     <Compile Include="Builders\TypedBuilder\ConstantEvaluator.cs" />
     <Compile Include="Builders\FieldsBuilder.cs" />
     <Compile Include="Builders\GeoNearOptionsBuilder.cs" />
